Fix incorrect dialogue in Cog Thief mini-game
For some reason the game says to use ALT or Delete to throw pies, but it's set to CTRL. I'm hoping John will either rectify this by fixing the mini-game to use Alt/Delete, or simply ensuring that Mac users can actually press CTRL to throw pies. I think that was an issue?